pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = ""
authors = ["Knt44kw <kento.yskw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.31.1"
google-cloud-vision = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
